Introduce GradleLintPluginTaskConfigurer and LegacyGradleLintPluginTaskConfigurer
